requirements.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.9.0
pyOpenSSL>=23.3.0
cryptography>=41.0.0

# Web crawling and browser automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# TLS and certificate handling
certifi>=2023.11.17

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0.1
toml>=0.10.2

# Reporting
jinja2>=3.1.2
# weasyprint>=60.0  # PDF generation (optional, requires C++ build tools on Windows)
markdown>=3.5.0

# Security scanning
python-whois>=0.8.0

# Statistical analysis
numpy>=1.26.0
scipy>=1.11.0

# CLI interface
click>=8.1.7
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# CI/CD
pygithub>=2.1.0  # For GitHub integration

# Utilities
python-dateutil>=2.8.2
colorama>=0.4.6

